Remove specialized postprocessors
