+++ app/services/field_correctors/__init__.py
@@ -1,9 +1,7 @@
 from .basic_cleaner import BasicFieldCorrector
-from .generic_cleaner import GenericFieldCleaner
 from .structured_cleaner import StructuredFieldCorrector
 
 __all__ = [
     "BasicFieldCorrector",
-    "GenericFieldCleaner",
     "StructuredFieldCorrector",
 ]
