--- app/services/postprocessors/form_postprocessor/__init__.py
@@ -1,3 +1,0 @@
-from .banorte_credito import BanorteCreditoPostProcessor
-
-__all__ = ["BanorteCreditoPostProcessor"]
