+++ tests/test_factories.py
@@ -2,9 +2,9 @@ import os
 import sys
 from pathlib import Path
 import pytest
+import types
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1] / "app"))
-import types
 sys.modules.setdefault("magic", types.SimpleNamespace(from_buffer=lambda *a, **k: "application/pdf"))
 
 class _DummyClient:
@@ -21,7 +21,7 @@ sys.modules.setdefault("aioboto3", types.SimpleNamespace(client=_dummy_client))
 sys.modules.setdefault("openai", types.ModuleType("openai"))
 sys.modules.setdefault(
     "transformers",
-    types.SimpleNamespace(pipeline=lambda *a, **k: lambda *args, **kwargs: [])
+    types.SimpleNamespace(pipeline=lambda *a, **k: lambda *args, **kwargs: []),
 )
 
 from services.ocr.factory import get_ocr_service
@@ -30,7 +30,7 @@ from services.ocr.textract.textract_ocr import AWSTextractOCRService
 from services.ai_refiners.gpt_refiner import GPTRefiner
 from services.ai_refiners.huggingface_refiner import HuggingFaceRefiner
 from services.postprocessors.postprocessor_factory import get_postprocessor
-from services.postprocessors.form_postprocessor.banorte_credito import BanorteCreditoPostProcessor
+from services.postprocessors.basic_postprocessor import BasicPostProcessor
 
 
 def test_get_ocr_service_default():
@@ -59,6 +59,7 @@ def test_get_ai_refiner_invalid():
         get_ai_refiner("invalid")
 
 
-def test_get_postprocessor_banorte_credito():
-    processor = get_postprocessor("banorte_credito")
-    assert isinstance(processor, BanorteCreditoPostProcessor)
+def test_get_postprocessor_default():
+    processor = get_postprocessor()
+    assert isinstance(processor, BasicPostProcessor)
+
